add exam schedule

+++ src/main/java/com/sep/vox/domain/model/exam/ExamPaperItem.java
@@ -1,5 +1,6 @@
 package com.sep.vox.domain.model.exam;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public class ExamPaperItem {
@@ -7,5 +8,5 @@ public class ExamPaperItem {
     private UUID sectionId;
     private UUID questionId;
     private int order;
-    private int weight;
+    private BigDecimal weight;
 }
